Die Werte der Managemet-Daten liegen, wenn möglich, zwischen 0 und 1. Vereinfacht das Ranking.
git-svn-id: svn://localhost/twospot_svn/trunk@823 8d73be03-56cb-46cc-8948-98b08ae92ef2

+++ Controller/src/org/prot/controller/management/services/IJmxResources.java
@@ -6,7 +6,7 @@ public interface IJmxResources
 {
 	public double loadAverage();
 
-	public long freeMemory();
+	public double freeMemory();
 
 	public double requestsPerSecond();
